+++ packages/client/src/WalletAPIClient.ts
@@ -28,6 +28,7 @@ import {
   WalletHandlers,
 } from "@ledgerhq/wallet-api-core";
 import { Storage } from "./modules/Storage";
+import { Bitcoin } from "./modules/Bitcoin";
 import { TransportWalletAPI } from "./TransportWalletAPI";
 
 const defaultLogger = new Logger("LL-PlatformSDK");
@@ -50,12 +51,18 @@ export class WalletAPIClient extends RpcNode<
 > {
   public storage: Storage;
 
+  /**
+   * Instance of the Bitcoin module
+   */
+  public bitcoin: Bitcoin;
+
   private logger: Logger;
 
   constructor(transport: Transport, logger: Logger = defaultLogger) {
     super(transport, requestHandlers);
     this.logger = logger;
     this.storage = new Storage(this);
+    this.bitcoin = new Bitcoin(this);
   }
 
   protected onRequest(request: RpcRequest) {
@@ -93,12 +100,8 @@ export class WalletAPIClient extends RpcNode<
       options,
     });
 
-    if ("error" in transactionSignResult) {
-      throw new RpcError(transactionSignResult.error);
-    }
-
     const safeResults = schemaTransactionSign.result.parse(
-      transactionSignResult.result
+      transactionSignResult
     );
 
     return Buffer.from(safeResults.signedTransactionHex, "hex");
@@ -127,12 +130,8 @@ export class WalletAPIClient extends RpcNode<
       }
     );
 
-    if ("error" in transactionSignResult) {
-      throw new RpcError(transactionSignResult.error);
-    }
-
     const safeResults = schemaTransactionSignAndBroadcast.result.parse(
-      transactionSignResult.result
+      transactionSignResult
     );
 
     return safeResults.transactionHash;
@@ -154,13 +153,7 @@ export class WalletAPIClient extends RpcNode<
       hexMessage: message.toString("hex"),
     });
 
-    if ("error" in messageSignResult) {
-      throw new RpcError(messageSignResult.error);
-    }
-
-    const safeResults = schemaMessageSign.result.parse(
-      messageSignResult.result
-    );
+    const safeResults = schemaMessageSign.result.parse(messageSignResult);
 
     return Buffer.from(safeResults.hexSignedMessage, "hex");
   }
@@ -183,13 +176,7 @@ export class WalletAPIClient extends RpcNode<
       currencyIds: params?.currencyIds,
     });
 
-    if ("error" in listAccountsResult) {
-      throw new RpcError(listAccountsResult.error);
-    }
-
-    const safeResults = schemaAccountList.result.parse(
-      listAccountsResult.result
-    );
+    const safeResults = schemaAccountList.result.parse(listAccountsResult);
 
     return safeResults.rawAccounts.map(deserializeAccount);
   }
@@ -212,12 +199,8 @@ export class WalletAPIClient extends RpcNode<
       currencyIds: params?.currencyIds,
     });
 
-    if ("error" in requestAccountsResult) {
-      throw new RpcError(requestAccountsResult.error);
-    }
-
     const safeResults = schemaAccountRequest.result.parse(
-      requestAccountsResult.result
+      requestAccountsResult
     );
 
     return deserializeAccount(safeResults.rawAccount);
@@ -235,12 +218,8 @@ export class WalletAPIClient extends RpcNode<
       accountId,
     });
 
-    if ("error" in receiveAccountsResult) {
-      throw new RpcError(receiveAccountsResult.error);
-    }
-
     const safeResults = schemaAccountReceive.result.parse(
-      receiveAccountsResult.result
+      receiveAccountsResult
     );
 
     return safeResults.address;
@@ -266,13 +245,7 @@ export class WalletAPIClient extends RpcNode<
       currencyIds: params?.currencyIds,
     });
 
-    if ("error" in listCurrenciesResult) {
-      throw new RpcError(listCurrenciesResult.error);
-    }
-
-    const safeResults = schemaCurrencyList.result.parse(
-      listCurrenciesResult.result
-    );
+    const safeResults = schemaCurrencyList.result.parse(listCurrenciesResult);
 
     return safeResults.currencies;
   }
@@ -293,12 +266,8 @@ export class WalletAPIClient extends RpcNode<
       params
     );
 
-    if ("error" in deviceTransportResult) {
-      throw new RpcError(deviceTransportResult.error);
-    }
-
     const safeResults = schemaDeviceTransport.result.parse(
-      deviceTransportResult.result
+      deviceTransportResult
     );
 
     return TransportWalletAPI.open({
@@ -321,12 +290,8 @@ export class WalletAPIClient extends RpcNode<
       {}
     );
 
-    if ("error" in walletCapabilitiesResult) {
-      throw new RpcError(walletCapabilitiesResult.error);
-    }
-
     const safeResults = schemaWalletCapabilities.result.parse(
-      walletCapabilitiesResult.result
+      walletCapabilitiesResult
     );
 
     return safeResults.methodIds;