+++ packages/core/src/spec/types/index.ts
@@ -11,3 +11,4 @@ export * from "./TransactionSignAndBroadcast";
 export * from "./WalletCapabilities";
 export * from "./StorageGet";
 export * from "./StorageSet";
+export * from "./BitcoinGetXPub";