+++ packages/server/src/index.ts
@@ -1,4 +1,2 @@
-export * from "rxjs";
-export * from "./errors";
 export * from "./types";
 export * from "./WalletAPIServer";