+++ packages/server/src/internalHandlers/device.ts
@@ -1,15 +1,16 @@
 import {
+  createNotImplementedByWallet,
   DeviceClose,
   DeviceExchange,
   DeviceTransport,
-  RpcError,
   schemaDeviceClose,
   schemaDeviceExchange,
   schemaDeviceTransport,
+  ServerError,
 } from "@ledgerhq/wallet-api-core";
-import { NOT_IMPLEMENTED_BY_WALLET } from "../errors";
 import type { RPCHandler } from "../types";
 
+const DEVICE_CLOSE = "device.close";
 export const close: RPCHandler<DeviceClose["result"]> = async (
   req,
   _,
@@ -17,10 +18,10 @@ export const close: RPCHandler<DeviceClose["result"]> = async (
 ) => {
   const safeParams = schemaDeviceClose.params.parse(req.params);
 
-  const walletHandler = handlers["device.close"];
+  const walletHandler = handlers[DEVICE_CLOSE];
 
   if (!walletHandler) {
-    throw new RpcError(NOT_IMPLEMENTED_BY_WALLET);
+    throw new ServerError(createNotImplementedByWallet(DEVICE_CLOSE));
   }
 
   const transportId = await walletHandler(safeParams);
@@ -30,6 +31,7 @@ export const close: RPCHandler<DeviceClose["result"]> = async (
   };
 };
 
+const DEVICE_EXCHANGE = "device.exchange";
 export const exchange: RPCHandler<DeviceExchange["result"]> = async (
   req,
   _,
@@ -37,10 +39,10 @@ export const exchange: RPCHandler<DeviceExchange["result"]> = async (
 ) => {
   const safeParams = schemaDeviceExchange.params.parse(req.params);
 
-  const walletHandler = handlers["device.exchange"];
+  const walletHandler = handlers[DEVICE_EXCHANGE];
 
   if (!walletHandler) {
-    throw new RpcError(NOT_IMPLEMENTED_BY_WALLET);
+    throw new ServerError(createNotImplementedByWallet(DEVICE_EXCHANGE));
   }
 
   const responseHex = await walletHandler(safeParams);
@@ -50,6 +52,7 @@ export const exchange: RPCHandler<DeviceExchange["result"]> = async (
   };
 };
 
+const DEVICE_TRANSPORT = "device.transport";
 export const transport: RPCHandler<DeviceTransport["result"]> = async (
   req,
   _,
@@ -57,10 +60,10 @@ export const transport: RPCHandler<DeviceTransport["result"]> = async (
 ) => {
   const safeParams = schemaDeviceTransport.params.parse(req.params);
 
-  const walletHandler = handlers["device.transport"];
+  const walletHandler = handlers[DEVICE_TRANSPORT];
 
   if (!walletHandler) {
-    throw new RpcError(NOT_IMPLEMENTED_BY_WALLET);
+    throw new ServerError(createNotImplementedByWallet(DEVICE_TRANSPORT));
   }
 
   const transportId = await walletHandler(safeParams);